// src/app/dashboard/interfaces/course.interface.ts
export interface Course {
  id?: number;
  nombre: string;
  modalidad: string;
  duracionMeses: string;
  costoMensual: number;
  costoCompleto: number;
  horarios: Schedule[];
  estado: string;
}

export interface Schedule {
  id?: number;
  turno: string;
  horaInicio: string;
  horaFin: string;
  estado: string;
}

export interface CourseDTO {
  id?: number;
  nombre: string;
  modalidad: string;
  costoMensual: number;
  costoCompleto: number;
  duracionMeses: number;
  estado: string;
}

export interface ScheduleDTO {
  id?: number;
  horaInicio: string;
  horaFin: string;
  turno: string;
  estado: string;
  cursoId: number;
}